Cleared DBLSTM
Uses the right train-test set partioning, because it turns out there is an official one.
Retrained the model, and also written a new zero padding, so that the collate function is a lot simpler,
or at least more is "hidden" from the user.
Introduced configargparse to prepare for different model setups.
Some introductory code for PPG frontend extraction is added.
Masking of loss function to show real results.