More faithful baseline added
Added BLSTM_paper.conf for a more faithful baseline.
Changed FFT parameters for a more faithful reproduction, this means that the FFT parameters need to be changed back for
the reproducibility.
Rewritten the model according to paper (2) in the new README.md
Added some scripts from spectrogram extraction and a temporary model for spectrogram based extraction which was part of an experiment that I made